+++ eslint.config.js
@@ -1,0 +1,16 @@
+import config from '@digitalbazaar/eslint-config/node-recommended';
+
+export default [
+  ...config,
+  {
+    files: [
+      'test/mocha/**/*.js'
+    ],
+    languageOptions: {
+      globals: {
+        // @bedrock/test global
+        assertNoError: true
+      }
+    }
+  }
+];